Move UserCommander.{removeEmail, makeEmailPrimary} to UserEmailAddressCommander
